+++ src/Problems_10_To_19/problem15.cpp
@@ -10,35 +10,35 @@ using std::map;
 
 struct coordinate {
     int x, y;
-    
+
     bool operator==(const coordinate &o) const {
         return x == o.x && y == o.y;
     }
-    
+
     bool operator<(const coordinate &o) const {
         return x < o.x || (x == o.x && y < o.y);
     }
 };
-map<coordinate, long> cache;
+map<coordinate, long long> cache;
 
-long numberOfPaths(int width, int height);
+long long numberOfPaths(int width, int height);
 
 const Problem15 Problem15::INSTANCE = Problem15();
 
 void Problem15::run() const {
     int gridSize = 20;
     cout << "Number Of Paths In 20 X 20 Grid From Top Left To Bottom Right Is: " << numberOfPaths(gridSize, gridSize) << endl << endl;
-    
+
 }
 
-long numberOfPaths(int width, int height) {
+long long numberOfPaths(int width, int height) {
     coordinate coord = {width, height};
-    
+
     if (cache.find(coord) != cache.end()) {
         return cache[coord];
     }
-    
-    long result = 0;
+
+    long long result = 0;
     if (width == 1 || height == 1) {
         result =  width + height;
     } else if (width == height) {
@@ -46,7 +46,7 @@ long numberOfPaths(int width, int height) {
     } else {
         result = numberOfPaths(width - 1, height) + numberOfPaths(width, height - 1);
     }
-    
+
     cache[coord] = result;
     return result;
 }
